+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Allow PHP 8.3
+- Allow symfony/console and symfony/event-dispatcher ^7.0.0
 
 ## [1.4.2](https://github.com/orisai/nette-console/compare/1.4.1...1.4.2) - 2024-03-12